+++ ExpandWorldMusic/ExpandWorldMusic.cs
@@ -11,9 +11,7 @@ public class EWM : BaseUnityPlugin
 {
   public const string GUID = "expand_world_music";
   public const string NAME = "Expand World Music";
-  public const string VERSION = "1.9";
-  public static string YamlDirectory = Path.Combine(Paths.ConfigPath, "expand_world");
-  public static string BackupDirectory = Path.Combine(Paths.ConfigPath, "expand_world_backups");
+  public const string VERSION = "1.10";
 
 #nullable disable
   public static CustomSyncedValue<string> valueMusicData;
@@ -37,6 +35,7 @@ public class EWM : BaseUnityPlugin
     try
     {
       Manager.SetupWatcher();
+      LocationManager.SetupWatcher();
     }
     catch (Exception e)
     {