+++ ExpandWorldMusic/Loader.cs
@@ -7,6 +7,7 @@ using UnityEngine;
 using UnityEngine.Networking;
 
 namespace ExpandWorld.Music;
+
 public class Loader
 {
   public static Dictionary<string, AudioClip> Clips = [];
@@ -68,7 +69,7 @@ public class Loader
     if (Clips.ContainsKey(path))
       return Clips[path];
     if (!File.Exists(path))
-      path = Path.Combine(EWM.YamlDirectory, path);
+      path = Path.Combine(Yaml.YamlDirectory, path);
     if (Clips.ContainsKey(path))
       return Clips[path];
     if (!File.Exists(path))