+++ ExpandWorldMusic/Manager.cs
@@ -9,7 +9,6 @@ namespace ExpandWorld.Music;
 public class Manager
 {
   public static string FileName = "expand_music.yaml";
-  public static string FilePath = Path.Combine(EWM.YamlDirectory, FileName);
   public static string Pattern = "expand_music*.yaml";
   public static List<MusicMan.NamedMusic> Originals = [];
 
@@ -18,9 +17,9 @@ public class Manager
   public static void ToFile()
   {
     if (!IsServer()) return;
-    if (File.Exists(FilePath)) return;
+    if (Yaml.Exists(FileName)) return;
     var yaml = Yaml.Write(MusicMan.instance.m_music.Select(Loader.ToData).ToList());
-    File.WriteAllText(FilePath, yaml);
+    Yaml.WriteFile(FileName, yaml);
   }
   public static void FromFile(string lines)
   {
@@ -73,11 +72,11 @@ public class Manager
     Log.Warning($"Adding {missing.Count} missing music to the expand_music.yaml file.");
     foreach (var item in missing)
       Log.Warning(item.m_name);
-    var yaml = File.ReadAllText(FilePath);
+    var yaml = Yaml.ReadFile(FileName);
     var data = Yaml.Write(missing.Select(Loader.ToData).ToList());
     // Directly appending is risky but necessary to keep comments, etc.
     yaml += "\n" + data;
-    File.WriteAllText(FilePath, yaml);
+    Yaml.WriteFile(FileName, yaml);
     return true;
   }
   private static void UpdateHashes()
@@ -93,7 +92,7 @@ public class Manager
   }
   public static void SetupWatcher()
   {
-    Watcher.Setup(EWM.YamlDirectory, EWM.BackupDirectory, Pattern, FromFile);
+    Yaml.Setup(Pattern, FromFile);
   }
 }
 
@@ -105,6 +104,6 @@ public class InitializeContent
   {
     Loader.InitializeDefaultClips();
     Manager.ToFile();
-    Manager.FromFile(Watcher.ReadFiles(EWM.YamlDirectory, Manager.Pattern));
+    Manager.FromFile(Yaml.ReadFiles(Manager.Pattern));
   }
 }