+++ publish/CHANGELOG.md
@@ -1,3 +1,7 @@
+- 1.10
+  - Adds a message to the `ew_musics` command to indicate that the available music clips were printed to the log file.
+  - Fixes config directory not being automatically created.
+
 - 1.9
   - Fixes the same location having multiple location entries causing multiple location music entries. Now only the first entry is used.
   - Fixes multiple same location music entries causing an error. Now only the first entry is used.